requirements.txt
joblib==1.2.0
numpy==1.21.6
pandas==1.3.5
PyYAML==6.0.1
scikit_learn==1.0.2
torch==1.12.1
tqdm==4.62.3
